feat: Export bit components
