+++ components/Avatar.js
@@ -5,7 +5,7 @@ import { color, border, space, layout } from 'styled-system';
 import themeGet from '@styled-system/theme-get';
 import { Flex } from '@rebass/grid';
 
-import { getCollectiveImage, getAvatarBorderRadius } from '../lib/utils';
+import { getCollectiveImage, getAvatarBorderRadius } from '../lib/image-utils';
 import { defaultImage, CollectiveType } from '../lib/constants/collectives';
 
 const getInitials = name => name.split(' ').reduce((result, value) => (result += value.slice(0, 1).toUpperCase()), '');
