+++ components/CollectiveCover.js
@@ -11,7 +11,8 @@ import momentTimezone from 'moment-timezone';
 import moment from 'moment';
 import { get } from 'lodash';
 import { withUser } from './UserProvider';
-import { prettyUrl, imagePreview } from '../lib/utils';
+import { prettyUrl } from '../lib/utils';
+import { imagePreview } from '../lib/image-utils';
 import Currency from './Currency';
 import Avatar from './Avatar';
 import Logo from './Logo';
