+++ components/Container.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { background, border, flexbox, shadow, color, position, layout, space, typography } from 'styled-system';
-import { cursor, clear, float, overflow, pointerEvents, whiteSpace } from '../lib/styled_system_custom';
+import { cursor, clear, float, overflow, pointerEvents, whiteSpace } from '../lib/styled-system-custom-properties';
 import propTypes from '@styled-system/prop-types';
 
 const Container = styled.div`
