+++ components/ListItem.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { space, textAlign, width, display, flexbox } from 'styled-system';
-import { listStyle } from '../lib/styled_system_custom';
+import { listStyle } from '../lib/styled-system-custom-properties';
 
 const ListItem = styled.li`
   list-style: none;
