+++ components/StyledLink.js
@@ -2,8 +2,8 @@ import PropTypes from 'prop-types';
 import styled, { css } from 'styled-components';
 import { border, color, layout, space, typography } from 'styled-system';
 import themeGet from '@styled-system/theme-get';
-import { whiteSpace, textDecoration } from '../lib/styled_system_custom';
-import { buttonSize, buttonStyle } from '../lib/theme';
+import { whiteSpace, textDecoration } from '../lib/styled-system-custom-properties';
+import { buttonSize, buttonStyle } from '../lib/theme/variants/button';
 
 /**
  * styled-component anchor tag using styled-system
@@ -13,6 +13,7 @@ import { buttonSize, buttonStyle } from '../lib/theme';
 const StyledLink = styled.a`
   color: ${themeGet('colors.primary.500')};
   cursor: pointer;
+  text-decoration: none;
 
   &:hover {
     color: ${themeGet('colors.primary.300')};
