+++ components/SubscriptionCard.js
@@ -7,7 +7,8 @@ import Currency from './Currency';
 import { get, cloneDeep } from 'lodash';
 import Link from './Link';
 import { getErrorFromGraphqlException } from '../lib/errors';
-import { firstSentence, getCurrencySymbol, imagePreview } from '../lib/utils';
+import { firstSentence, getCurrencySymbol } from '../lib/utils';
+import { imagePreview } from '../lib/image-utils';
 import { getStripe } from '../lib/stripe';
 import { defaultBackgroundImage } from '../lib/constants/collectives';
 import colors from '../lib/constants/colors';
