+++ components/Text.js
@@ -1,7 +1,7 @@
 import styled, { css } from 'styled-components';
 import { color, display, space, typography } from 'styled-system';
 
-import { textTransform, whiteSpace, wordBreak, cursor } from '../lib/styled_system_custom';
+import { textTransform, whiteSpace, wordBreak, cursor } from '../lib/styled-system-custom-properties';
 
 export const P = styled.p.attrs(props => ({
   // Overrides default margin Y to avoid global styles
