+++ components/contribution-flow/index.js
@@ -16,7 +16,8 @@ import { AmountTypes } from '../../lib/constants/tiers-types';
 import { VAT_OPTIONS } from '../../lib/constants/vat';
 import { Router } from '../../server/pages';
 import { getStripe, stripeTokenToPaymentMethod } from '../../lib/stripe';
-import { compose, formatCurrency, getEnvVar, parseToBoolean, reportValidityHTML5 } from '../../lib/utils';
+import { getEnvVar } from '../../lib/env-utils';
+import { compose, formatCurrency, parseToBoolean, reportValidityHTML5 } from '../../lib/utils';
 import { getPaypal } from '../../lib/paypal';
 import { getRecaptcha, getRecaptchaSiteKey, unloadRecaptcha } from '../../lib/recaptcha';
 import { addCreateCollectiveMutation } from '../../lib/graphql/mutations';
