+++ lib/google-maps.js
@@ -1,6 +1,7 @@
 import { get } from 'lodash';
 
-import { getEnvVar, loadScriptAsync } from './utils';
+import { loadScriptAsync } from './utils';
+import { getEnvVar } from './env-utils';
 
 const getGoogleMapsScriptUrl = () => {
   const apiKey = getEnvVar('GOOGLE_MAPS_API_KEY');
