+++ lib/recaptcha.js
@@ -1,4 +1,5 @@
-import { getEnvVar, loadScriptAsync } from './utils';
+import { loadScriptAsync } from './utils';
+import { getEnvVar } from './env-utils';
 
 const getRecaptchaSiteKey = () => getEnvVar('RECAPTCHA_SITE_KEY');
