+++ lib/stripe.js
@@ -1,6 +1,7 @@
 import Payment from 'payment';
 
-import { getEnvVar, loadScriptAsync } from './utils';
+import { loadScriptAsync } from './utils';
+import { getEnvVar } from './env-utils';
 
 const stripeInstances = {};
