+++ lib/theme/index.js
@@ -1,4 +1,3 @@
-import { variant } from 'styled-system';
 import { transparentize } from 'polished';
 
 // Import defaults
@@ -239,22 +238,6 @@ export const generateTheme = ({
   sizes: {
     navbarHeight: 68,
   },
-});
-
-// Export variants
-export const buttonStyle = variant({
-  key: 'buttons',
-  prop: 'buttonStyle',
-});
-
-export const buttonSize = variant({
-  key: 'buttonSizes',
-  prop: 'buttonSize',
-});
-
-export const messageType = variant({
-  key: 'messageTypes',
-  prop: 'type',
 });
 
 // Generate and export main theme
