+++ lib/theme/variants/button.js
@@ -1,0 +1,11 @@
+import { variant } from 'styled-system';
+
+export const buttonStyle = variant({
+  key: 'buttons',
+  prop: 'buttonStyle',
+});
+
+export const buttonSize = variant({
+  key: 'buttonSizes',
+  prop: 'buttonSize',
+});
