+++ lib/theme/variants/message.js
@@ -1,0 +1,6 @@
+import { variant } from 'styled-system';
+
+export const messageType = variant({
+  key: 'messageTypes',
+  prop: 'type',
+});
