+++ lib/utils.js
@@ -1,4 +1,3 @@
-import { get } from 'lodash';
 import getSymbolFromCurrency from 'currency-symbol-map';
 
 import loadScript from 'load-script';
@@ -100,9 +99,6 @@ export const getPrecisionFromAmount = amount => {
   return amount.toString().slice(-2) === '00' ? 0 : CurrencyPrecision.DEFAULT;
 };
 
-/** Retrieve variables set in the environment */
-export const getEnvVar = v => (process.browser ? get(window, ['__NEXT_DATA__', 'env', v]) : get(process, ['env', v]));
-
 export function parseToBoolean(value) {
   let lowerValue = value;
   // check whether it's string
@@ -113,93 +109,6 @@ export function parseToBoolean(value) {
     return true;
   }
   return false;
-}
-
-export const getBaseImagesUrl = () => getEnvVar('IMAGES_URL');
-
-export function resizeImage(imageUrl, { width, height, query, baseUrl }) {
-  if (!imageUrl) {
-    return null;
-  }
-  if (imageUrl.substr(0, 1) === '/') {
-    return imageUrl;
-  } // if image is a local image, we don't resize it with the proxy.
-  if (imageUrl.substr(0, 4).toLowerCase() !== 'http') {
-    return null;
-  } // Invalid imageUrl;
-  if (!query && imageUrl.match(/\.svg$/)) {
-    return imageUrl;
-  } // if we don't need to transform the image, no need to proxy it.
-  let queryurl = '';
-  if (query) {
-    queryurl = encodeURIComponent(query);
-  } else {
-    if (width) {
-      queryurl += `&width=${width}`;
-    }
-    if (height) {
-      queryurl += `&height=${height}`;
-    }
-  }
-
-  return `${getBaseImagesUrl() || baseUrl || ''}/proxy/images?src=${encodeURIComponent(imageUrl)}${queryurl}`;
-}
-
-export function isValidImageUrl(src) {
-  return src && (src.substr(0, 1) === '/' || src.substr(0, 4).toLowerCase() === 'http');
-}
-
-export function imagePreview(src, defaultImage, options = { width: 640 }) {
-  if (typeof options.width === 'string') {
-    options.width = Number(options.width.replace(/rem/, '')) * 10;
-  }
-  if (typeof options.height === 'string') {
-    options.height = Number(options.height.replace(/rem/, '')) * 10;
-  }
-
-  if (src) {
-    return resizeImage(src, options);
-  }
-  if (isValidImageUrl(defaultImage)) {
-    return defaultImage;
-  }
-  return null;
-}
-
-export function getAvatarBorderRadius(collectiveType) {
-  return collectiveType === 'USER' ? '50%' : '25%';
-}
-
-function createCollectiveImageUrl(collective, params = {}) {
-  const sections = [getBaseImagesUrl(), collective.slug];
-
-  sections.push(params.name || 'avatar');
-
-  for (const key of ['style', 'height', 'width']) {
-    if (params[key]) {
-      sections.push(params[key]);
-    }
-  }
-
-  return `${sections.join('/')}.${params.format || 'png'}`;
-}
-
-export function getCollectiveImage(collective, params = {}) {
-  // If available use the imageUrl provided by the API
-  if (collective.imageUrl) {
-    return collective.imageUrl;
-  }
-
-  return createCollectiveImageUrl(collective, { ...params });
-}
-
-export function getCollectiveBackgroundImage(collective, params = {}) {
-  // If available use the imageUrl provided by the API
-  if (collective.backgroundImageUrl) {
-    return collective.backgroundImageUrl;
-  }
-
-  return createCollectiveImageUrl(collective, { ...params, name: 'background' });
 }
 
 export function prettyUrl(url) {
